Replace manual parse by regex
Set has a logic error which raises a bad_format_exception if the plus
sign is located before the dash in a version string.
This can happen if the pre-release part is given and the build part con-
tains hypens which is a valid semantic version string.

This is fixed using a regex to match the parts of the semantic version
and setting them seperatelly using set().

This is a braking change as is requires a compiler/standard library
which supports the <regex> header.